+++ Tests/PrimitiveTests/UInt64Tests.swift
@@ -13,7 +13,7 @@ class RAWUInt64Tests: XCTestCase {
     
     func testInitWithRAWData() {
         let bytes: [UInt8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01]
-        let value = UInt64(RAW_size: UInt64(MemoryLayout<UInt64>.size), RAW_data:bytes)
+        let value = UInt64(RAW_size:MemoryLayout<UInt64>.size, RAW_data:bytes)
         XCTAssertEqual(value, 1)
     }
 }
